// src/features/editableProfileCard/model/types/editableProfileCardSchema.ts
import { type Profile } from '@/entities/Profile/model/types/profile'
import { type ValidateProfileError } from '../consts/consts'

export interface ProfileSchema {
  data?: Profile
  form?: Profile
  isLoading?: boolean
  error?: string
  readonly?: boolean
  validateErrors?: ValidateProfileError[]
}